Create RandomPasswordGenerator.java
This is a program that generates a
random password for the user. Prompt the user to
enter the desired length of the password and
specify whether it should include numbers,
lowercase letters, uppercase letters, and special
characters. Then it will generate the password accordingly
and display it to the user.